+++ kernel/syscalls/syscalls.c
@@ -47,6 +47,7 @@ typedef syscall_word_t (*syscall_func_t)(
 static syscall_func_t syscall_table[MAX_SYSCALLS] = {
     [__NR_exit] = (syscall_func_t)sys_exit,
     [__NR_fork] = (syscall_func_t)sys_fork,
+    [__NR_clone] = (syscall_func_t)sys_clone,
     [__NR_read] = (syscall_func_t)sys_read,
     [__NR_write] = (syscall_func_t)sys_write,
     [__NR_open] = (syscall_func_t)sys_open,
@@ -64,6 +65,7 @@ static syscall_func_t syscall_table[MAX_SYSCALLS] = {
     [__NR_pause] = (syscall_func_t)sys_pause,
     [__NR_utime] = (syscall_func_t)sys_utime,
     [__NR_getpid] = (syscall_func_t)sys_getpid,
+    [__NR_gettid] = (syscall_func_t)sys_gettid,
     [__NR_getppid] = (syscall_func_t)sys_getppid,
     [__NR_setuid] = (syscall_func_t)sys_setuid,
     [__NR_getuid] = (syscall_func_t)sys_getuid,
@@ -164,6 +166,7 @@ static syscall_func_t syscall_table[MAX_SYSCALLS] = {
     [__NR_recvfrom]  = (syscall_func_t)sys_recvfrom,
     [__NR_socket_shutdown] = (syscall_func_t)sys_socket_shutdown,
     [__NR_resolve]   = (syscall_func_t)sys_resolve,
+    [__NR_thread_exit] = (syscall_func_t)sys_thread_exit,
 
 };
 
@@ -684,6 +687,10 @@ void sys_exit(int status)
         return;
     }
 
+    if (proc->type == TASK_TYPE_THREAD) {
+        sys_thread_exit(status);
+    }
+
     if (proc->type != TASK_TYPE_PROCESS || !proc->process) {
         KERROR("[EXIT] sys_exit called from non-process task! Name = %s\n", proc->name);
         return;
@@ -1161,21 +1168,24 @@ int syscall_handler(uint32_t syscall_num, uint32_t arg1, uint32_t arg2,
 int sys_getpid(void)
 {
     task_t *proc = task_current_local();
+    process_t *process = task_get_process(proc);
 
-    if (proc && proc->type == TASK_TYPE_PROCESS && proc->process) {
-        return proc->process->pid;
-    }
-    return 0;
+    return process ? process->pid : 0;
+}
+
+int sys_gettid(void)
+{
+    task_t *task = task_current_local();
+
+    return task ? (int)task->task_id : 0;
 }
 
 int sys_getppid(void)
 {
     task_t *proc = task_current_local();
+    process_t *process = task_get_process(proc);
 
-    if (proc && proc->type == TASK_TYPE_PROCESS && proc->process) {
-        return proc->process->ppid;
-    }
-    return 0;
+    return process ? process->ppid : 0;
 }
 
 int sys_getuid(void)